feat: refactor login handling and introduce AuthenticateAndIssueToken for token management
