+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/crediya/api/user/register/RegisterHandler.java
@@ -2,7 +2,7 @@ package co.com.crediya.api.user.register;
 
 import co.com.crediya.api.user.register.dto.mapper.UserMapperDto;
 import co.com.crediya.api.user.register.dto.request.CreateUserRequestDto;
-import co.com.crediya.api.user.support.RegisterUserAssembler;
+import co.com.crediya.api.user.register.support.RegisterUserAssembler;
 import co.com.crediya.api.user.support.ResponseUtils;
 import co.com.crediya.api.user.validations.CreateUserValidator;
 import co.com.crediya.usecase.user.IUserUseCase;
