+++ infrastructure/entry-points/reactive-web/src/main/java/co/com/crediya/api/user/register/support/RegisterUserAssembler.java
@@ -1,4 +1,4 @@
-package co.com.crediya.api.user.support;
+package co.com.crediya.api.user.register.support;
 
 import co.com.crediya.api.user.register.dto.request.CreateUserRequestDto;
 import co.com.crediya.model.role.Role;
